frontend/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
mysql-connector-python==8.1.0
python-socketio==5.8.0
python-engineio==4.7.1
eventlet==0.33.3
